+++ app/client/src/constants/messages.ts
@@ -177,3 +177,8 @@ export const GOOGLE_RECAPTCHA_DOMAIN_ERROR =
 export const SERVER_API_TIMEOUT_ERROR =
   "Appsmith server is taking too long to respond. Please try again after some time";
 export const DEFAULT_ERROR_MESSAGE = "There was an unexpected error";
+
+export const OAUTH_AUTHORIZATION_SUCCESSFUL =
+  "OAuth Authorization was successful!";
+export const OAUTH_AUTHORIZATION_FAILED =
+  "OAuth Authorization failed. Please check the credentials and try again.";
